+++ parameters.tf
@@ -57,6 +57,14 @@ resource "aws_ssm_parameter" "db_host" {
   tags      = module.labels.tags
 }
 
+resource "aws_ssm_parameter" "db_pool_size" {
+  overwrite = true
+  name      = "${local.config_var_prefix}db_pool_size"
+  type      = "String"
+  value     = var.db_pool_size
+  tags      = module.labels.tags
+}
+
 resource "aws_ssm_parameter" "db_port" {
   overwrite = true
   name      = "${local.config_var_prefix}db_port"
@@ -94,6 +102,14 @@ resource "aws_ssm_parameter" "default_region" {
   name      = "${local.config_var_prefix}default_region"
   type      = "String"
   value     = var.default_region
+  tags      = module.labels.tags
+}
+
+resource "aws_ssm_parameter" "disable_valid_key_check" {
+  overwrite = true
+  name      = "${local.config_var_prefix}disable_valid_key_check"
+  type      = "String"
+  value     = var.disable_valid_key_check
   tags      = module.labels.tags
 }
 
@@ -137,6 +153,14 @@ resource "aws_ssm_parameter" "certificate_audience" {
   tags      = module.labels.tags
 }
 
+resource "aws_ssm_parameter" "hsts_max_age" {
+  overwrite = true
+  name      = "${local.config_var_prefix}hsts_max_age"
+  type      = "String"
+  value     = var.hsts_max_age
+  tags      = module.labels.tags
+}
+
 resource "aws_ssm_parameter" "jwt_issuer" {
   overwrite = true
   name      = "${local.config_var_prefix}jwt_issuer"
@@ -166,6 +190,14 @@ resource "aws_ssm_parameter" "native_regions" {
   name      = "${local.config_var_prefix}native_regions"
   type      = "String"
   value     = var.native_regions
+  tags      = module.labels.tags
+}
+
+resource "aws_ssm_parameter" "onset_date_mandatory" {
+  overwrite = true
+  name      = "${local.config_var_prefix}onset_date_mandatory"
+  type      = "String"
+  value     = var.onset_date_mandatory
   tags      = module.labels.tags
 }
 
@@ -214,6 +246,14 @@ resource "aws_ssm_parameter" "security_code_lifetime_mins" {
   name      = "${local.config_var_prefix}security_code_lifetime_mins"
   type      = "String"
   value     = var.code_lifetime_mins
+  tags      = module.labels.tags
+}
+
+resource "aws_ssm_parameter" "security_code_removal_mins" {
+  overwrite = true
+  name      = "${local.config_var_prefix}security_code_removal_mins"
+  type      = "String"
+  value     = var.code_removal_mins
   tags      = module.labels.tags
 }
 
@@ -285,7 +325,15 @@ resource "aws_ssm_parameter" "time_zone" {
   overwrite = true
   name      = "${local.config_var_prefix}time_zone"
   type      = "String"
-  value     = var.sms_template
+  value     = var.time_zone
+  tags      = module.labels.tags
+}
+
+resource "aws_ssm_parameter" "upload_max_keys" {
+  overwrite = true
+  name      = "${local.config_var_prefix}upload_max_keys"
+  type      = "String"
+  value     = var.upload_max_keys
   tags      = module.labels.tags
 }
 
@@ -293,7 +341,7 @@ resource "aws_ssm_parameter" "upload_token_lifetime_mins" {
   overwrite = true
   name      = "${local.config_var_prefix}upload_token_lifetime_mins"
   type      = "String"
-  value     = var.refresh_token_expiry
+  value     = var.upload_token_lifetime_mins
   tags      = module.labels.tags
 }
 
@@ -302,6 +350,14 @@ resource "aws_ssm_parameter" "use_test_date_as_onset_date" {
   name      = "${local.config_var_prefix}use_test_date_as_onset_date"
   type      = "String"
   value     = var.use_test_date_as_onset_date
+  tags      = module.labels.tags
+}
+
+resource "aws_ssm_parameter" "variance_offset_mins" {
+  overwrite = true
+  name      = "${local.config_var_prefix}variance_offset_mins"
+  type      = "String"
+  value     = var.variance_offset_mins
   tags      = module.labels.tags
 }