+++ variables.tf
@@ -37,6 +37,10 @@ variable "api_gateway_account_creation_enabled" {
   description = "APIGateway account creation flag, this is used for CloudWatch logging, should only have one of these per account/region, this flag allows disabling if one already exists"
   default     = true
 }
+variable "api_gateway_minimum_compression_size" {
+  description = "APIGateway minimum response size to compress for the REST API. Integer between -1 and 10485760 (10MB). Setting a value greater than -1 will enable compression, -1 disables compression (default)"
+  default     = -1
+}
 variable "api_gateway_throttling_burst_limit" {
   description = "APIGateway throttling burst limit, default is -1 which does not enforce a limit"
   default     = -1
@@ -57,6 +61,11 @@ variable "bastion_asg_desired_count" {
 variable "bastion_enabled" {
   description = "Bastion enabled, does not provision the bastion, only allows using a bastion, see the bastion_asg_desired_count variable"
   default     = true
+}
+variable "bastion_instance_type" {
+  description = "Bastion EC2 instance type"
+  type        = string
+  default     = "t2.small"
 }
 
 # #########################################
@@ -95,6 +104,13 @@ variable "push_eu_certificate_arn" {
 variable "default_ecr_max_image_count" {
   description = "Default ECR image retention count used for purging the ECR repositories"
   default     = 30
+}
+# #########################################
+# Load Balancer
+# #########################################
+variable "lb_push_ssl_policy" {
+  description = "Name of TLS policy in use"
+  default     = "ELBSecurityPolicy-TLS-1-2-Ext-2018-06"
 }
 
 # #########################################
@@ -179,10 +195,23 @@ variable "enable_sms_publishing_with_aws" {
   description = "Enable sending SMS via a SNS topic"
   default     = false
 }
+variable "sms_delivery_status_success_sampling_rate" {
+  description = "Percentage sampling of delivery logs sent into CloudWatch"
+  default     = 0
+}
+variable "sms_monthly_spend_limit" {
+  description = "Monthly limit for SMS"
+  default     = 100 # Note: this value has to be requested to the AWS Support as a quota increase ticket.
+}
 
 # #########################################
 # WAF
 # #########################################
+variable "attach_waf" {
+  description = "Attach WAF to ALBs and API Gateway - Sometimes need to detach for pen testing"
+  default     = true
+  type        = bool
+}
 # List of allowed country alpha 2 codes, see https://www.iso.org/obp/ui/#search
 # If this is empty then we do not restrict based on country
 variable "waf_geo_allowed_countries" {
@@ -215,6 +244,14 @@ variable "api_ecs_autoscale_max_instances" {
 }
 variable "api_ecs_autoscale_min_instances" {
   description = "ECS API service ASG min count"
+  default     = 1
+}
+variable "api_ecs_autoscale_scale_down_adjustment" {
+  description = "ECS API service ASG scaling scale down adjustment"
+  default     = -1
+}
+variable "api_ecs_autoscale_scale_up_adjustment" {
+  description = "ECS API service ASG scaling scale up adjustment"
   default     = 1
 }
 variable "api_image_tag" {
@@ -276,6 +313,10 @@ variable "code_length" {
 variable "code_lifetime_mins" {
   description = "Lifetime in minutes of the one-time upload codes"
 }
+variable "code_removal_mins" {
+  description = "Lifetime in minutes before a one-time upload code is removed from the database"
+  default     = "10080"
+}
 variable "daily_registrations_reporter_email_subject" {
   description = "daily-registrations-reporter lambda email subject text"
   default     = ""
@@ -284,6 +325,10 @@ variable "daily_registrations_reporter_schedule" {
   description = "daily-registrations-reporter lambda CloudWatch schedule"
   default     = ""
 }
+variable "db_pool_size" {
+  description = "Maximum number of clients the db pool should contain"
+  default     = "30"
+}
 variable "default_country_code" {
   description = "Default ISO country code to use for parsing mobile numbers provided to push service"
   default     = ""
@@ -291,6 +336,10 @@ variable "default_country_code" {
 variable "default_region" {
   description = "Default region to use for exposure key uploads where the region is not provided"
   default     = ""
+}
+variable "disable_valid_key_check" {
+  description = "Flag to disable whether exposure keys which are still valid are ignored when generating export files"
+  default     = "false"
 }
 variable "download_schedule" {
   description = "download lambda CloudWatch schedule"
@@ -342,6 +391,10 @@ variable "health_check_timeout" {
 variable "health_check_unhealthy_threshold" {
   description = "Health check unhealthy threshold for ALB health checks"
   default     = 2
+}
+variable "hsts_max_age" {
+  description = "The time, in seconds, that the browser should remember that a site is only to be accessed using HTTPS."
+  default     = "300" // 5 minutes
 }
 variable "lambda_authorizer_memory_size" {
   description = "authorizer lambda memory size"
@@ -425,7 +478,7 @@ variable "lambda_download_timeout" {
 }
 variable "lambda_exposures_memory_size" {
   description = "exposures lambda memory size"
-  default     = 128
+  default     = 256
 }
 variable "lambda_exposures_s3_key" {
   description = "exposures lambda S3 key if using - file path"
@@ -433,7 +486,7 @@ variable "lambda_exposures_s3_key" {
 }
 variable "lambda_exposures_timeout" {
   description = "exposures lambda timeout"
-  default     = 15
+  default     = 60
 }
 variable "lambda_provisioned_concurrencies" {
   description = "Map of lambdas to use provisioned concurrency i.e. { \"authorizer\" : 300 }"
@@ -521,6 +574,10 @@ variable "migrations_image_tag" {
   description = "Image tag for the ECS Migrations container"
   default     = "latest"
 }
+variable "onset_date_mandatory" {
+  description = "Flag whether onsetDate/symptomDate is mandatory"
+  default     = "false"
+}
 variable "optional_lambdas_to_include" {
   description = "List of optional lambdas to include"
   default     = []
@@ -555,6 +612,14 @@ variable "push_ecs_autoscale_max_instances" {
 }
 variable "push_ecs_autoscale_min_instances" {
   description = "ECS Push service ASG min count"
+  default     = 1
+}
+variable "push_ecs_autoscale_scale_down_adjustment" {
+  description = "ECS Push service ASG scaling scale down adjustment"
+  default     = -1
+}
+variable "push_ecs_autoscale_scale_up_adjustment" {
+  description = "ECS Push service ASG scaling scale up adjustment"
   default     = 1
 }
 variable "push_image_tag" {
@@ -608,6 +673,10 @@ variable "sms_template" {
   description = "SMS message template"
   default     = ""
 }
+variable "sms_type" {
+  description = "SMS message type"
+  default     = "Transactional"
+}
 variable "symptom_date_offset" {
   description = "Offset in hours subtracted from the symptom or onset date for uploads"
   default     = "0"
@@ -617,6 +686,11 @@ variable "time_zone" {
   default     = "UTC"
 }
 variable "token_lifetime_mins" {
+  description = "Token lifetime in minutes"
+}
+variable "upload_max_keys" {
+  description = "Maximum keys accepted in a single upload request"
+  default     = "15"
 }
 variable "upload_schedule" {
   description = "upload lambda CloudWatch schedule"
@@ -629,6 +703,10 @@ variable "upload_token_lifetime_mins" {
 variable "use_test_date_as_onset_date" {
   description = "Flag to use the testDate as the onsetDate if the latter is omitted"
   default     = "false"
+}
+variable "variance_offset_mins" {
+  description = "Variance offset in minutes to add to lifetime of keys to check if they are still valid"
+  default     = "120"
 }
 variable "verify_rate_limit_secs" {
   description = "Time in seconds a user must wait before attempting to verify a one-time upload code"